requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
httpx
sounddevice
numpy
scipy
